Getting rid of compile warnings (mostly types, and some uninitialized vars)

+++ pslr_enum.h
@@ -30,6 +30,7 @@
 #define PSLR_ENUM_H
 
 #include <stdint.h>
+#include <ctype.h>
 
 typedef enum {
     PSLR_COLOR_SPACE_SRGB,
@@ -198,7 +199,7 @@ typedef enum {
     PSLR_SCENE_MODE_MAX
 } pslr_scene_mode_t;
 
-int str_comparison_i (const char *s1, const char *s2, int n);
+int str_comparison_i (const char *s1, const char *s2, size_t n);
 int find_in_array( const char** array, int length, char* str );
 
 pslr_color_space_t get_pslr_color_space( char *str );
